Finishing up  Course review activity, Not able to integrate develop for some reason?

+++ UMHUB/app/src/main/java/comp3350/umhub/presentation/WriteCourseReviewActivity.java
@@ -57,7 +57,7 @@ public class WriteCourseReviewActivity extends AppCompatActivity {
 
     public void buttonSubmitCourseReviewOnClick(View view){
         String review = reviewEditText.getText().toString();
-        int reviewScore = Double.parseDouble(reviewScoreEditText.getText().toString());
+        int reviewScore = Integer.parseInt(reviewScoreEditText.getText().toString());
 
         CourseReview courseReview = new CourseReview(currentUser,course,reviewScore,review);
         accessCourseReviews.addReview(courseReview);
